Started working on SSA, entering SSA should be done, keeping in mind that phi's are represented, but no retargeting for phi results will occur until leaving SSA, this is because it breaks SSA form, so be aware that uses after phi's have not been retargetted yet, should probably add a linearization type for phi's temporarily for partial retargetting to a phi, to better track instructions not yet actually retargetted between SSA enter and leave

+++ Proton.VM/IR/IRAssembly.cs
@@ -72,7 +72,7 @@ namespace Proton.VM.IR
                         {
                             IRLocal local = new IRLocal(this);
                             local.ParentMethod = method;
-                            local.Index = (uint)method.Locals.Count;
+                            local.Index = method.Locals.Count;
                             method.Locals.Add(local);
                         }
                     }
